+++ packages/core/app-layout/sandbox/components/icons/AppGruceLogo.vue
@@ -1,0 +1,31 @@
+<template>
+  <svg
+    v-bind="$attrs"
+    fill="none"
+    height="26"
+    width="29"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      clip-rule="evenodd"
+      d="m9.144 21.802.693-.894h5.126l2.672 3.415-.475 1.159h-6.61l.158-1.159-1.564-2.52Z"
+      fill="#169FCC"
+      fill-rule="evenodd"
+    /><path
+      clip-rule="evenodd"
+      d="m10.711 10.158 2.476-4.278h2.884L29 20.903l-1.003 4.579h-5.543l.347-1.286-12.09-14.038Z"
+      fill="#14B59A"
+      fill-rule="evenodd"
+    /><path
+      clip-rule="evenodd"
+      d="m13.717 5.068 1.185-2.193L18.457 0l6.093 4.778-.79.807 1.06 1.468v1.572l-3.036 2.482-5.094-6.04h-2.973Z"
+      fill="#1BC263"
+      fill-rule="evenodd"
+    /><path
+      clip-rule="evenodd"
+      d="M4.301 14.621h1.61l4.2-3.514 5.565 6.469-1.57 2.356H8.97l-3.547 4.512-.815 1.038H0v-5.53l4.301-5.33Z"
+      fill="#16BDCC"
+      fill-rule="evenodd"
+    />
+  </svg>
+</template>