+++ packages/core/app-layout/sandbox/components/icons/index.ts
@@ -1,0 +1,2 @@
+export { default as AppGruceLogo } from './AppGruceLogo.vue'
+export { default as AppLogo } from './AppLogo.vue'