+++ packages/core/cli/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/Kong/public-ui-components/compare/@kong-ui/core-public-cli@0.1.7...@kong-ui/core-public-cli@0.2.0) (2023-01-23)
+
+
+### ✨ Features
+
+* **core-app-layout:** create package ([#30](https://github.com/Kong/public-ui-components/issues/30)) ([cc60e80](https://github.com/Kong/public-ui-components/commit/cc60e80d4fa44b21013d8d533e624cded01ef8fa))
+
+
+### ♻️ Chores
+
+* **cli:** add types directory ([0a7fafa](https://github.com/Kong/public-ui-components/commit/0a7fafa76cefe5e99d2ff4a8bcd740c6003ed0f7))
+
+
+
+
+
 ## [0.1.7](https://github.com/Kong/public-ui-components/compare/@kong-ui/core-public-cli@0.1.6...@kong-ui/core-public-cli@0.1.7) (2023-01-23)